apps/backend/audits/tests/requirements-test.txt
# Testing dependencies for audit feature
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
